+++ internal/sinks/sink.go
@@ -1,3 +1,4 @@
+// Package sinks contains sinks for emitting ValidationResults.
 package sinks
 
 import (
@@ -11,13 +12,16 @@ import (
 	"github.com/validator-labs/validator/pkg/types"
 )
 
-var SinkEmissionFailed = errors.New("sink emission failed")
+// ErrSinkEmissionFailed is returned when emitting a validation result to a sink fails.
+var ErrSinkEmissionFailed = errors.New("sink emission failed")
 
+// Sink is an interface for sending validation results to a sink.
 type Sink interface {
 	Configure(c Client, config map[string][]byte) error
 	Emit(result v1alpha1.ValidationResult) error
 }
 
+// NewSink returns a new Sink based on the provided SinkType.
 func NewSink(sinkType types.SinkType, log logr.Logger) Sink {
 	switch sinkType {
 	case types.SinkTypeAlertmanager:
@@ -29,10 +33,12 @@ func NewSink(sinkType types.SinkType, log logr.Logger) Sink {
 	}
 }
 
+// Client is an HTTP client for a Sink.
 type Client struct {
 	hclient *http.Client
 }
 
+// NewClient returns a new Client with the provided timeout.
 func NewClient(timeout time.Duration) *Client {
 	client := &http.Client{
 		Timeout: timeout,
