+++ pkg/constants/constants.go
@@ -1,6 +1,10 @@
+// Package constants contains validator constants.
 package constants
 
 const (
+	// ValidationRulePrefix is the prefix for all validation rules.
 	ValidationRulePrefix string = "validation"
-	ValidatorConfig      string = "validator-config"
+
+	// ValidatorConfig is the name of the default validator config.
+	ValidatorConfig string = "validator-config"
 )
